% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityHelpers.R
\name{getCensusYears}
\alias{getCensusYears}
\title{getCensusYears returns a list of community names 
returns the list of census years contained with in 
the data frame that is passes in}
\usage{
getCensusYears(censusDF)
}
\arguments{
\item{cencusDF}{dataframe with census data}
}
\description{
getCensusYears returns a list of community names 
returns the list of census years contained with in 
the data frame that is passes in
}
\examples{
getCensusYears (censusDF )
getCensusYears (df)

}

