% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityHelpers.R
\name{getCommunityClasses}
\alias{getCommunityClasses}
\title{getCommunityClasses returns a list of community names 
returns the list of community classes contained with in 
the data frame that is passes in}
\usage{
getCommunityClasses(censusDF)
}
\arguments{
\item{cencusDF}{dataframe with census data}
}
\description{
getCommunityClasses returns a list of community names 
returns the list of community classes contained with in 
the data frame that is passes in
}
\examples{
getCommunityClasses (censusDF )
getCommunityClasses (df)
}

