% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityHelpers.R
\name{getCommunityCodes}
\alias{getCommunityCodes}
\title{getCommunityCodes returns a list of community codes 
returns the list of community codes contained with in 
the data frame that is passes in}
\usage{
getCommunityCodes(censusDF)
}
\arguments{
\item{cencusDF}{dataframe with census data}
}
\description{
getCommunityCodes returns a list of community codes 
returns the list of community codes contained with in 
the data frame that is passes in
}
\examples{
getCommunityCodes (censusDF )
getCommunityCodes (df)
}

