% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityHelpers.R
\name{getCommunityNames}
\alias{getCommunityNames}
\title{getCommunityNames returns a list of community names 
returns the list of community codes contained with in 
the data frame that is passes in}
\usage{
getCommunityNames(censusDF)
}
\arguments{
\item{cencusDF}{dataframe with census data}
}
\description{
getCommunityNames returns a list of community names 
returns the list of community codes contained with in 
the data frame that is passes in
}
\examples{
getCommunityNames (censusDF )
getCommunityNames (df)
}

