% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initCommData}
\alias{initCommData}
\title{initCommData Retreive and initalize census data}
\usage{
initCommData()
}
\arguments{
\item{path}{path to directory containing the csv files}

\item{csvpattern}{mathing pattern for the csv files}
}
\description{
This function will load data from multiple csv files into a single data frame.
}
\examples{
myDF <- initCommData()
initCommData()
}

