% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initWardData}
\alias{initWardData}
\title{initWardData Retreive and initalize census data
This function will load data from multiple csv files into a single data frame.}
\usage{
initWardData()
}
\arguments{
\item{path}{path to directory containing the csv files}

\item{csvpattern}{mathing pattern for the csv files}
}
\description{
initWardData Retreive and initalize census data
This function will load data from multiple csv files into a single data frame.
}
\examples{
myDF <- initCommData()
initWardData()
}

