% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityHelpers.R
\name{toCommCode}
\alias{toCommCode}
\title{toCommCode Covert Community Name to Short Code}
\usage{
toCommCode(censusDF, commName)
}
\arguments{
\item{commName}{string containing the community name}

\item{df}{dataframe with census data}
}
\description{
toCommCode Covert Community Name to Short Code
}
\examples{
toCommCode (censusDF,"Highland Park")
toCommCode (myDF,commName)
}

