% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityHelpers.R
\name{toLongName}
\alias{toLongName}
\title{toLongName Convert three letter community code to full name}
\usage{
toLongName(myDF, code)
}
\arguments{
\item{code}{three letter community code}

\item{df}{dataframe with census data}
}
\description{
toLongName Convert three letter community code to full name
}
\examples{
toLongName (censusDF,"HPK")
toLongName(myDF,core)
}

