% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communityHelpers.R
\name{toLongName}
\alias{toLongName}
\title{Convert three letter community code to full name}
\usage{
toLongName(myDF, code)
}
\arguments{
\item{code}{three letter community code}

\item{df}{dataframe with census data}
}
\description{
This function will load data from multiple csv files into a single data frame.
}
\examples{
toLongName(myDF,core)
}

